Extract Rust gatekeeper metrics module
